requirements.txt
asn1crypto==0.24.0
astroid==2.1.0
autopep8==1.4.3
certifi==2019.3.9
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.5
Django==2.1.5
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-filter==2.1.0
django-silk==3.0.1
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
drf-yasg==1.14.0
future==0.17.1
gprof2dot==2016.10.13
idna==2.8
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==5.4.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
pycodestyle==2.5.0
pycparser==2.19
Pygments==2.3.1
PyJWT==1.7.1
pylint==2.2.2
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
ruamel.yaml==0.15.89
Serializer==0.2.1
serializers==0.2.4
simplejson==3.16.0
six==1.12.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.24.1
uWSGI==2.0.17.1
wrapt==1.11.1
